# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Arbiter Inference Routing Gateway
  version: 0.1.0
  description: OpenAI-compatible contract for routed inference and cost attribution.
paths:
  /v1/chat/completions:
    post:
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Non-streaming completion response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '402':
          description: Request exceeds its hard cost budget.
        '422':
          description: Invalid request.
components:
  schemas:
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: auto or a pinned tier such as tier:premium.
        messages:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [role, content]
            properties:
              role:
                type: string
                enum: [system, user, assistant, tool]
              content:
                type: string
        stream:
          type: boolean
          default: false
        x_arbiter:
          type: object
          additionalProperties: false
          properties:
            max_cost_usd:
              type: number
              minimum: 0
            min_quality_tier:
              type: string
              enum: [standard, high, critical]
            cache:
              type: string
              enum: [allow, bypass, refresh]
            tenant_id:
              type: string
              minLength: 1
    ChatCompletionResponse:
      type: object
      required: [object, id, model, choices, usage]
      properties:
        object:
          type: string
          example: chat.completion
        id:
          type: string
        model:
          type: string
        choices:
          type: array
          minItems: 1
          items:
            type: object
            required: [index, message, finish_reason]
            properties:
              index:
                type: integer
              message:
                type: object
                required: [role, content]
                properties:
                  role:
                    type: string
                    enum: [assistant]
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          required: [prompt_tokens, completion_tokens, total_tokens]
          properties:
            prompt_tokens:
              type: integer
              minimum: 0
            completion_tokens:
              type: integer
              minimum: 0
            total_tokens:
              type: integer
              minimum: 0